--- ObjectPrinting/Solved/ObjectPrinter.cs
@@ -1,10 +1,0 @@
-namespace ObjectPrinting.Solved
-{
-    public class ObjectPrinter
-    {
-        public static PrintingConfig<T> For<T>()
-        {
-            return new PrintingConfig<T>();
-        }
-    }
-}
