+++ ObjectPrinting_Should/Person.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ObjectPrinting.Solved.Tests
+namespace ObjectPrinting_Should
 {
     public class Person
     {
